feat(history): merge balance cards into one tabbed card
Weekly Balance and Training Mix stacked to two full cards on History.
Combine them into a single Balance card with Patterns / Training Mix
tabs, tighten row spacing (min-h 44->40px, gap 8->4px), and gate each
axis' fetch behind its flag via a new enabled option on the debt hooks.
