--- src/components/ModalityBalance/index.ts
@@ -1,2 +1,0 @@
-export * from './ModalityBalance';
-export * from './ModalityBalanceContainer';
