+++ src/components/TrainingBalance/index.ts
@@ -1,0 +1,2 @@
+export * from './TrainingBalance';
+export * from './TrainingBalanceContainer';
