--- src/components/WeeklyBalance/index.ts
@@ -1,2 +1,0 @@
-export * from './WeeklyBalance';
-export * from './WeeklyBalanceContainer';
