+++ src/components/index.ts
@@ -4,7 +4,6 @@ export * from './KettlebellGlyph';
 export * from './Loading';
 export * from './LoadingScreen';
 export * from './PageLoading';
-export * from './ModalityBalance';
 export * from './ModifyCountButtons';
 export * from './OverflowMenu';
 export * from './PWAInstallPrompt';
@@ -13,10 +12,10 @@ export * from './PremiumGate';
 export * from './ProgramRecommendationPreviewDialog';
 export * from './ProgramTags';
 export * from './SafeAreaWrapper';
+export * from './TrainingBalance';
 export * from './Sidebar';
 export * from './SpotifyMiniPlayer';
 export * from './StackFitNote';
 export * from './TrialStatusPill';
 export * from './ValueCarousel';
-export * from './WeeklyBalance';
 export * from './WeightUnitTabs';
